cleanup tests and imports
