+++ exec/exec.go
@@ -9,14 +9,10 @@
 package exec
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"os"
 	"os/exec"
-
-	"github.com/openconfig/kne/logshim"
-	log "k8s.io/klog/v2"
 )
 
 // A Cmd is an interface representing a command to run.
@@ -60,60 +56,4 @@ func (c command) Run() error {
 		return fmt.Errorf("%q failed: %v", c.cmd.String(), err)
 	}
 	return nil
-}
-
-// runCommand is a wrapper utility function that creates and runs a command with
-// various inputs and settings.
-func runCommand(writeLogs bool, in []byte, cmd string, args ...string) ([]byte, error) {
-	c := Command(cmd, args...)
-	var out bytes.Buffer
-	c.SetStdout(&out)
-	c.SetStderr(&out)
-	if writeLogs {
-		outLog := logshim.New(func(v ...interface{}) {
-			log.Info(append([]interface{}{"(" + cmd + "): "}, v...)...)
-		})
-		errLog := logshim.New(func(v ...interface{}) {
-			log.Warning(append([]interface{}{"(" + cmd + "): "}, v...)...)
-		})
-		defer func() {
-			outLog.Close()
-			errLog.Close()
-		}()
-		c.SetStdout(io.MultiWriter(outLog, &out))
-		c.SetStderr(io.MultiWriter(errLog, &out))
-	}
-	if len(in) > 0 {
-		c.SetStdin(bytes.NewReader(in))
-	}
-	err := c.Run()
-	return out.Bytes(), err
-}
-
-// LogCommand runs the specified command but records standard output
-// with log.Info and standard error with log.Warning.
-func LogCommand(cmd string, args ...string) error {
-	_, err := runCommand(true, nil, cmd, args...)
-	return err
-}
-
-// LogCommandWithInput runs the specified command but records standard output
-// with log.Info and standard error with log.Warning. in is sent to
-// the standard input of the command.
-func LogCommandWithInput(in []byte, cmd string, args ...string) error {
-	_, err := runCommand(true, in, cmd, args...)
-	return err
-}
-
-// OutLogCommand runs the specified command but records standard output
-// with log.Info and standard error with log.Warning. Standard output
-// and standard error are also returned.
-func OutLogCommand(cmd string, args ...string) ([]byte, error) {
-	return runCommand(true, nil, cmd, args...)
-}
-
-// OutCommand runs the specified command and returns any standard output
-// as well as any errors.
-func OutCommand(cmd string, args ...string) ([]byte, error) {
-	return runCommand(false, nil, cmd, args...)
 }
