+++ topo/node/gobgp/gobgp_test.go
@@ -3,12 +3,11 @@ package gobgp
 import (
 	"testing"
 
-	"google.golang.org/protobuf/encoding/prototext"
-	"google.golang.org/protobuf/proto"
-
-	"github.com/h-fam/errdiff"
+	"github.com/openconfig/gnmi/errdiff"
 	topopb "github.com/openconfig/kne/proto/topo"
 	"github.com/openconfig/kne/topo/node"
+	"google.golang.org/protobuf/encoding/prototext"
+	"google.golang.org/protobuf/proto"
 )
 
 func TestNew(t *testing.T) {
