+++ topo/node/node_test.go
@@ -8,14 +8,13 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/h-fam/errdiff"
+	"github.com/openconfig/gnmi/errdiff"
+	topopb "github.com/openconfig/kne/proto/topo"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	kfake "k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/rest"
-
-	topopb "github.com/openconfig/kne/proto/topo"
 )
 
 func NewNR(impl *Impl) (Node, error) {
